sql-agent-service/requirements.txt
# ── Core web / API ─────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.0

# ── Database ────────────────────────────────────────────────────────────────
psycopg2-binary==2.9.9

# ── HTTP client (AI API calls) ───────────────────────────────────────────────
requests==2.31.0

# ── Environment / Config ─────────────────────────────────────────────────────
python-dotenv==1.0.1

# ── Optional: alternative ORM / DB toolkit ───────────────────────────────────
# Uncomment if you prefer SQLAlchemy for query building:
# sqlalchemy==2.0.30
